Support for EXR images, dry-run option to see what would be downloaded

+++ run.py
@@ -29,13 +29,23 @@ def getImagesForCategory(url):
         yield image_name
 
 
-def downloadImage(name, quality):
+def downloadImage(name, quality='1k', outdir='.', dry_run=False):
     url = BASE_URL + "/files/hdris/%s_%s.hdr" % (name, quality)
-    response = requests.get(url, stream=True)
-    path = os.path.join(outdir, name + ".hdr")
-    with open(path, "wb") as handle:
-        for data in response.iter_content():
-            handle.write(data)
+    if requests.head(url).status_code != 200:
+        url = BASE_URL + "/files/hdris/%s_%s.exr" % (name, quality)
+        if requests.head(url).status_code != 200:
+             return None  # Failure
+       
+    
+    path = os.path.join(outdir, os.path.basename(url))
+    if os.path.isfile(path):
+        print(f"**Skipping {name}, output file already exists")
+    if not dry_run:
+        response = requests.get(url, stream=True)
+        with open(path, "wb") as handle:
+            for data in response.iter_content():
+                handle.write(data)
+    return path
 
 
 parser = argparse.ArgumentParser(description="A script to download all HDRIs of a certain category")
@@ -43,12 +53,19 @@ parser.add_argument("--mode", choices=['list', 'download'], help="Action to pefo
 parser.add_argument("--category_url", help="Category url (use '--mode list' to see options)", default="/hdris/category/?c=all&o=popular")
 parser.add_argument("--quality", choices=["1k", "2k", "4k", "8k", "16k"], help="Quality of HDRI to download", default="8k")
 parser.add_argument("--outdir", type=str, help="Output folder", default=".")
-cmd = parser.parse_args()
+parser.add_argument("--dry-run", action="store_true", help="Do not download -- only list images")
+args = parser.parse_args()
 
+print(args)
 
-if cmd.mode == "list":
+if args.mode == "list":
     print("\n".join("%s: %s" % (name, url) for url, name in getCategories()))
 else:
     os.makedirs(args.outdir, exist_ok=True)
-    for name in tqdm(getImagesForCategory(cmd.category_url)):
-        downloadImage(name, cmd.quality, args.outdir)
+    names = list(getImagesForCategory(args.category_url))
+    for name in tqdm(names):
+        path = downloadImage(name, args.quality, args.outdir, args.dry_run)
+        if path is None:
+            tqdm.write(f"** Bad URl: {name}")
+        else:
+            tqdm.write(path)
